requirements.txt
pydantic-ai
python-dotenv>=1.0.0
google-generativeai>=0.5.0
asyncio
pydantic>=2.0.0